// engemedical-worker/src/soc/types/IResultadoExameDataFicha.ts
export type ResultadoDataFichaExameRequest = {
  empresa?: string;
  funcionario?: string;
  dataInicio?: string;
  dataFim?: string;
};

export type ResultadoDataFichaExame = {
  EMPRESA: string;
  NOME: string;
  CODFUNCIONARIO: string;
  NOMEFUNCIONARIO: string;
  MATRICULA: string;
  DATAFICHA: string;
  TIPOFICHA: string;
  DATAEXAME: string;
  CODEXAME: string;
  NOMEEXAME: string;
  SEQUENCIARESULTADOEXAME: string; // Adicionar no pedido de exame
  SEQUENCIAFICHA: string;
  RESULTADOEXAME: string;
};